TimeCard_py/requirements.txt
brotlipy==0.7.0
certifi==2020.12.5
cffi @ file:///C:/ci/cffi_1613247279197/work
chardet @ file:///C:/ci/chardet_1607690654534/work
configparser @ file:///tmp/build/80754af9/configparser_1602256123427/work
cryptography @ file:///C:/ci/cryptography_1613401566183/work
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome @ file:///C:/ci/pycryptodome_1613765534229/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///C:/ci/pysocks_1605287845585/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
six @ file:///C:/ci/six_1605187374963/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1611694770489/work
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
